Modify the components
