+++ client/src/components/index.js
@@ -1,0 +1,17 @@
+import BeUpdated from './BeUpdated';
+import Cart from './Cart';
+import Categories from './Categories';
+import FeaturedProducts from './FeaturedProducts';
+import Footer from './Footer';
+import Header from './Header';
+import MainImage from './MainImage';
+import Menu from './Menu';
+import OurJournal from './OurJournal';
+import OurStory from './OurStory';
+import OurStoryAboutUs from './OurStoryAboutUs';
+import OurStoryNavigationMenu from './OurStoryNavigationMenu';
+import OurStoryOurMission from './OurStoryOurMission';
+import OurStoryProduction from './OurStoryProduction';
+import OneFeatured from './OneFeatured';
+
+export {BeUpdated, Cart, Categories, FeaturedProducts, Footer, Header, MainImage, Menu, OurJournal, OurStory, OurStoryAboutUs, OurStoryNavigationMenu, OneFeatured, OurStoryOurMission, OurStoryProduction};
