+++ client/src/pages/JournalPage/index.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
-import Footer from '../components/Footer';
+import Footer from '../../components/Footer';
+import './style.css';
 
 class JournalPage extends Component {
   render() {
