+++ client/src/pages/JournalPage/style.css
@@ -1,0 +1,139 @@
+/************************* Journal Page ***************************************/
+.journal-page {
+  width: 84.5%;
+  margin: 0 auto;
+  margin-bottom: 150px;
+  padding-top: 50px;
+}
+
+.journal-page section {
+  width: 730px;
+  padding: 0 15px;
+  margin: 0 0 60px 197px;
+  }
+
+  .journal-page section p, .journal-page section h1 {
+  text-align: left;
+  font-family: Montserrat,Verdana, Geneva, sans-serif;
+  text-align: inherit;
+  line-height: 24px;
+  font-weight: 300;
+  margin: 0 0 24px;
+  word-wrap: break-word;
+  box-sizing: border-box;
+  font-style: normal;
+  transition: opacity 0.24s ease-in-out;
+  font-size: 14px;
+}
+
+.journal-page section h1 strong {
+  font-weight: 400;
+}
+
+.journal-page div {
+  background: #f7f7f7;
+  display: inline-block;
+  width: 100%;
+  padding: 30px 0;
+  margin: 0 auto;
+  text-align: center;
+  font-weight: 300;
+  box-sizing: border-box;
+  height: 89px;
+  vertical-align: middle;
+}
+
+.journal-page div a {
+  margin-right: 10px;
+}
+
+.journal-page div a i {
+  font-size: 18px;
+}
+
+.journal-page div a:nth-child(1) i {
+  color: #00aced;
+}
+
+.journal-page div a:nth-child(2) i {
+  color: #3b5998;
+}
+
+.journal-page div a:nth-child(3) i {
+  color: #dd4b39;
+}
+
+.journal-page div a:nth-child(4) i {
+  color: #d32027;
+}
+
+.journal-page div a:nth-child(5) i {
+  color: #1985c9;
+}
+
+
+
+/*************************** Media Queries ************************************/
+@media only screen and (max-width: 1024px) {
+  .journal-page {
+    width: 85.8826171875%;
+    margin: 0 auto;
+  }
+  .journal-page section {
+    margin: 0 0 60px 170px;
+  }
+  .journal-page div {
+    margin-bottom: 50px;
+  }
+
+}
+
+@media only screen and (max-width: 768px) {
+  .journal-page {
+    width: 672.094px;
+    margin: 0 auto;
+    float: none;
+  }
+  .journal-page div {
+    width: 100%;
+  }
+  .journal-page section {
+    padding: 0;
+    margin: 0;
+    width: 100%;
+    margin-bottom: 50px;
+  }
+  .journal-page section p, .journal-page section h1 {
+    font-size: 12px;
+  }
+  .all-watches-page {
+    width: 87.4453%;
+  }
+}
+
+@media only screen and (max-width: 425px) {
+  .journal-page {
+    width: 394.25px;
+  }
+  .journal-page div {
+    float: none;
+  }
+  .journal-page div a i {
+    font-size: 25px;
+  }
+  .journal-page div a {
+    margin-right: 20px;
+  }
+}
+
+@media only screen and (max-width: 375px) {
+  .journal-page {
+    width: 353.75px;
+  }
+}
+
+@media only screen and (max-width: 320px) {
+  .journal-page {
+    width: 309.188px;
+  }
+}
