+++ client/src/pages/OurStory/index.js
@@ -1,10 +1,5 @@
 import React, { Component } from 'react';
-import OurStoryAboutUs from '../components/OurStoryAboutUs';
-import OurStoryOurMission from '../components/OurStoryOurMission';
-import OurStoryProduction from '../components/OurStoryProduction';
-import OurStoryNavigationMenu from '../components/OurStoryNavigationMenu';
-import Footer from '../components/Footer';
-import './style.css';
+import { OurStoryAboutUs, OurStoryOurMission, OurStoryProduction, OurStoryNavigationMenu, Footer } from '../../components';
 
 class OurStory extends Component {
   render() {
